+++ doc/conf.py
@@ -10,7 +10,7 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
-"""Placeholder docstring"""
+"""Configuration for generating readthedocs docstrings."""
 from __future__ import absolute_import
 
 import pkg_resources
@@ -97,4 +97,6 @@ autosectionlabel_prefix_document = True
 
 
 def setup(app):
+    sys.stdout.write("Generating JumpStart model table...")
+    sys.stdout.flush()
     create_jumpstart_model_table()