+++ src/sagemaker/fw_utils.py
@@ -74,6 +74,11 @@ SM_DATAPARALLEL_SUPPORTED_FRAMEWORK_VERSIONS = {
         "2.6",
         "2.6.0",
         "2.6.2",
+        "2.6.3",
+        "2.7",
+        "2.7.1",
+        "2.8",
+        "2.8.0",
     ],
     "pytorch": [
         "1.6",